refactor: split pseudo-Japanese helpers
