+++ src/helpers/meditationPage.js
@@ -2,7 +2,7 @@
  * Initialize the Meditation page.
  *
  * @pseudocode
- * 1. Import `setupLanguageToggle` from `pseudoJapanese.js` and `loadQuote` from `quoteBuilder.js`.
+ * 1. Import `setupLanguageToggle` from `pseudoJapanese/ui.js` and `loadQuote` from `quoteBuilder.js`.
  * 2. Define `setupMeditationPage` which:
  *    a. Retrieves the quote element from the DOM.
  *    b. Calls `loadQuote` to display a random quote.
@@ -12,7 +12,7 @@
  *
  * @returns {void}
  */
-import { setupLanguageToggle } from "./pseudoJapanese.js";
+import { setupLanguageToggle } from "./pseudoJapanese/ui.js";
 import { onDomReady } from "./domReady.js";
 import { initTooltips } from "./tooltip.js";
 import { loadQuote } from "./quoteBuilder.js";
